backend/ai-sytem-backend/pyproject.toml
[project]
name = "ai-sytem-backend"
version = "0.1.0"
description = ""
authors = [{ name = "CJPD00", email = "carlosjavierpacheco60@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "selenium (>=4.31.0,<5.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "python-logging-loki (>=0.3.1,<0.4.0)",
]

[tool.poetry]
packages = [{ include = "ai_sytem_backend", from = "src" }]

[tool.poetry.scripts]
start="src.ai_sytem_backend.main:start"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
